Successfully intergrated multiple scripts
